chore(postcss): externalize sass config
