+++ packages/css-dev-tools/postcssPluginConfig.js
@@ -14,36 +14,9 @@ const MOZAIC_ENV = process.env.MOZAIC_ENV
 const mozaicEnvScssVar =
   MOZAIC_ENV === "production" ? MOZAIC_ENV : "development"
 
-// test for user configured additional paths
-const additionalPaths = CM.getKey("sass.includePaths")
-
-// test for user configured custom tokens paths
-const tokensBuildPath = CM.getKey("tokens.localTokensExportPath")
-
-const basePaths = [
-  "./node_modules/@mozaic-ds/styles/",
-  "./node_modules/@mozaic-ds/styles/components/",
-  "./node_modules/@mozaic-ds/styles/generic/",
-  "./node_modules/@mozaic-ds/styles/layouts/",
-  "./node_modules/@mozaic-ds/styles/settings-tools/",
-  "./node_modules/@mozaic-ds/styles/typography/",
-  "./node_modules/@mozaic-ds/styles/utilities/",
-  "./node_modules/",
-]
-
-const tokensPath = tokensBuildPath
-  ? basePaths.concat([`${tokensBuildPath}scss/`])
-  : basePaths.concat(["./node_modules/@mozaic-ds/tokens/build/scss/"])
-
-const includePaths = additionalPaths
-  ? tokensPath.concat(additionalPaths)
-  : tokensPath
-
-// test for user configured space indent
-const userIndent = CM.getKey("sass.indentWidth")
-const indentWidth = userIndent ? userIndent : 2
-
 const styleLintConfig = require("./styleLintConfig")
+
+const sassConfig = require("./sassConfig")
 
 // load browserlist config
 const borwserslistConfig = CM.getKey("browserslist")
@@ -56,11 +29,7 @@ const plugins = [
   stylelint({ config: styleLintConfig }),
   reporter({ clearReportedMessages: true }),
   cssprepend(`$mozaic-env: ${mozaicEnvScssVar};`),
-  sass({
-    includePaths,
-    outputStyle: "expanded",
-    indentWidth,
-  }),
+  sass(sassConfig),
   base64({
     pattern: /<svg.*<\/svg>/i,
     prepend: "data:image/svg+xml;base64,",
@@ -75,11 +44,7 @@ const plugins = [
 
 const productionPlugins = [
   cssprepend(`$mozaic-env: ${mozaicEnvScssVar};`),
-  sass({
-    includePaths,
-    outputStyle: "expanded",
-    indentWidth,
-  }),
+  sass(sassConfig),
   base64({
     pattern: /<svg.*<\/svg>/i,
     prepend: "data:image/svg+xml;base64,",
